Move createHistory function to Router folder.
Imagine a world where I had not forgotten to do this when I moved all of
the other integrations.
